Fixed a subtle bug and changed a file name.
Fixed a sublte bug that was causing an ArrayIndexBoundExceptions in GaussJordan2.java (due to the program trying to access an invalid index of the Matrix). 

I improved the main method in both Gauss.java and GaussJordan2.java.

Renamed GaussJordan2.java to GaussJordan.java.

I originally named the file GaussJordan2 because the whole project was in a different directory, in which GaussJordan.java already existed, which was a barely written attempt to code a Gauss Jordan program on a matrix of floats. Later, I moved the project into its folder, but kept the name.